+++ app/app/src/main/java/com/admt/inventoryTracker/AddStockRequestParameters.java
@@ -6,7 +6,7 @@ public class AddStockRequestParameters {
     public String Barcode;
     public String ItemId;
     public String LocationId;
-    public Integer BulkItemCount; // only applicable for bulk items
+    public Double BulkItemCount; // only applicable for bulk items
     public Double ItemQuantityToAdd; // allows for adding a partial pack to the system
     public String ExpiryDate;
